[IMP] fs_field: Improve table management
